alarms:  decouple remote log processing from local logging context
Testing and debugging has revealed several issues arising when the alarms
server is sharing a domain with other components.  The most crucial is
that the LogEntryAppender is added to the root appender tree.  This means
that the appending logic is carried out synchronously by the local calling thread,
differently from the case where the server is remote.  Since the work done on
this thread includes potential regex matching (of an indefinite number of patterns)
and storing through the DataNucleus layer to the database, this can have considerable
impact on the performance of the domain.

The most immediate solution would be to thread the handling of events in
the appender.  But an even better solution would be not to detach the remote
appender and substitute it with the LogEntryAppender for this domain, but
to uncouple instead the handling of incoming events sent on the socket from
the logging context altogether.

This has been achieved by modifying the server to do the work of the
SocketServer in logback, but without context-related configuration.
A special SocketNode is used which passes the incoming event directly to
the LogEventAppender, now renamed LogEventHandler, without sending it
through the local logging mechanism.  Concomitantly, the handler is not
attached to any logger and hence does not process local events.

Several other modifications have been made in support of more efficient
handling of remote logging events.  Here is a detailed list of changes:

1.  Used SimpleSocketServer as model for processing connections and embedded
    this directly into LogEntryServer.

2.  Adapted logback SocketNode to handle the event object by passing it
    directly to the the LogEntryHandler.

3.  Refactored LogEntryAppender into LogEntryHandler.  It no longer
    extends the Appender base.  The handler uses an executor to process
    the event if the event level is >= root level.  The two other
    logback appenders are directly activated by the task, as before.
    [While asynchronous handling here is desirable, but multiple workers
    may not be necessary; this currently defaults to 1, but further
    testing is required to find the optimal setting.]

4.  There is a new option (default is false) to store ONLY alarms to the
    permanent store.

5.  The LogEntryServerSocketNode derives the originating host from the
    client InetAddress. This, along with the proprietary alarms properties
    for domain and service, are set on the logging event mdc before passing
    the event object to the handler.

6.  As a consequence, the RemoteMDCFilter has been removed from remote logger.
    References to the domain and cell originally pointing to constants from
    this class have by replaced by the standard CDC constants.

7.  The “source host” property now being redundant, the related CLI option
    has been removed from the shell (with man page modified to reflect this),
    and from the admin interface.

8.  Two new properties have been added to the alarms.properties and reflected
    in the batch checks:

    alarms.service.limits.number-of-workers=1
    (one-of?true|false)alarms.db.alarms-only=false

9.  The remote appender (renamed to socket) has been wrapped with an async appender.
    I assume for the moment that the default settings are sufficient.